pyproject.toml
[tool.poetry]
name = "physicAIn"
version = "0.1.0"
description = "A medical physician software powered using AI."
authors = [
    "Ahmed Sabry <ahmedmsabry2000@gmail.com>",
    "Youssef Mahmoud <>",
    "Zeyad Ganainy <>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
flask = "2.2.2"
flask-restful = "0.3.9"
matplotlib = "3.3.4"
numpy = "1.24.0"
pandas = "1.5.2"
scikit-learn = "1.2.0"
Pillow = "9.3.0"
nltk = "3.8.0"
pre-commit = "2.20.0"
black = "22.3.0"
requests = "2.28.1"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
